+++ software/reachy/changelog.md
@@ -8,3 +8,4 @@
 ## Version 1.1.0
 
 * Add support for sensors in WsIO (eg. for Unity simulation)
+* Add support for dual camera (left or right), only one can be active at the same time.